samples/python/agents/sql_agent/pyproject.toml
[project]
name = "a2a-sample-agent-sql"
version = "0.1.0"
description = "Sample SQL Agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "a2a-samples",
    "python-dotenv>=1.1.0",
    "mysql-connector-python",
    "pandas",
    "pydantic>=1.10",
    "click",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-samples = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"